Match regex with word boundaries
This will make sure we won't treat part of word as a match.
E.g. without word boundaries DTS for audio matched also TS for quiality.

Fix test expected from TS to BRRip where needed.
Add some output to the unittest so it will be easy to read on error.

Signed-off-by: Roi Dayan <roi.dayan@gmail.com>