Additional Quality and Codec Possibilities
Added 720p and 1080p to quality.
Added space as an option to period for H.264, X.265 to be represented as H 264, X 265.  I personally don't like it, but it's now common in torrent names.
Added HEVC to codecs, as it's a technical alternate name for h.265, and popular in torrent names.